Refactor: ERD 변경에 따른 엔티티 수정, 엔티티 수정에 따른 코드 수정

+++ src/main/java/com/cruise/parkinglotto/domain/WeightSectionStatistics.java
@@ -21,10 +21,10 @@ public class WeightSectionStatistics extends BaseEntity {
     private WeightSection section;
 
     @Column(nullable = false)
-    private Long applicantsCount;
+    private Integer applicantsCount;
 
     @Column(nullable = false)
-    private Long winnerCount;
+    private Integer winnerCount;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "draw_id")
